requirements.txt
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
certifi==2018.4.16
cffi==1.11.5
constantly==15.1.0
cryptography==2.2.2
cssselect==1.0.3
elasticsearch==6.2.0
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
lxml==4.2.1
parsel==1.4.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
PyDispatcher==2.0.5
pyOpenSSL==17.5.0
python-dotenv==0.8.2
queuelib==1.5.0
Scrapy==1.5.0
service-identity==17.0.0
six==1.11.0
Twisted==17.9.0
Unidecode==1.0.22
urllib3==1.22
uuid==1.30
w3lib==1.19.0
zope.interface==4.5.0
